+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - StockRules: Fix tracking 'userIdealQty' to 'useIdealQty'
 - COPY OF PRODUCT : do not copy lastproductionprice value.
 - Invoice : add header and footer height parameter on PrintSetting object
+- FileTool : Use try-with-resources for "BufferedReader"
 
 ## [5.1.11] - 2019-11-05
 ## Improvements